Improved if-else logic around if language standard is recognized or not

+++ src/example.py
@@ -14,7 +14,7 @@ if (__name__ == '__main__'):
     clean_up_build_directories: bool = True
     """
 
-    Arithmetic_codebase: CodeBase
+    Arithmetic_codebase: CodeBase | None = None
 
     try:
         """
@@ -36,7 +36,7 @@ if (__name__ == '__main__'):
         Arithmetic_codebase = \
             CodeBase('Arithmetic',
                      Path.cwd()/f'example_C_code',
-                     language_standard='C 2011')
+                     language_standard='C 2018')
         Arithmetic_codebase.generate_as_executable()
 
     except Exception:
@@ -58,5 +58,6 @@ if (__name__ == '__main__'):
                     shutil.rmtree(present_arithmetic_codebase.build_directory)
         """
 
-        if Arithmetic_codebase.build_directory.exists():
-            shutil.rmtree(Arithmetic_codebase.build_directory)
+        if Arithmetic_codebase:
+            if Arithmetic_codebase.build_directory.exists():
+                shutil.rmtree(Arithmetic_codebase.build_directory)
